+++ Example/Tests/Tests.m
@@ -9,6 +9,7 @@
 // https://github.com/Specta/Specta
 
 #import "SEGPayloadBuilder.h"
+#import <MixpanelGroup.h>
 
 SpecBegin(InitialSpecs)
 
@@ -16,15 +17,20 @@ describe(@"Mixpanel Integration", ^{
     __block SEGMixpanelIntegration *integration;
     __block Mixpanel *mixpanel;
     __block MixpanelPeople *mixpanelPeople;
+    __block MixpanelGroup *mixpanelGroup;
 
     beforeEach(^{
         mixpanel = mock([Mixpanel class]);
         mixpanelPeople = mock([MixpanelPeople class]);
         [given([mixpanel people]) willReturn:mixpanelPeople];
 
+        mixpanelGroup = mock([MixpanelGroup class]);
+
         integration = [[SEGMixpanelIntegration alloc] initWithSettings:@{
             @"trackAllPages" : @1,
-            @"setAllTraitsByDefault" : @1
+            @"setAllTraitsByDefault" : @1,
+            @"enableEuropeanUnionEndpoint" : @1,
+            @"groupIdentifierTraits" : @[@"group_id", @"group_name", @"group_idtest2"]
         } andMixpanel:mixpanel];
     });
 
@@ -57,7 +63,7 @@ describe(@"Mixpanel Integration", ^{
         }];
         [verify(mixpanelPeople) trackCharge:@19.99];
     });
-    
+
     it(@"track with property increments", ^{
         [integration setSettings:@{
             @"people" : @1,
@@ -69,20 +75,20 @@ describe(@"Mixpanel Integration", ^{
         [verify(mixpanel) track:@"Purchased Item" properties:@{
             @"revenue" : @19.99
         }];
-    });    
+    });
 
     it(@"screen", ^{
         [integration screen:[SEGPayloadBuilder screen:@"Home"]];
 
         [verify(mixpanel) track:@"Viewed Home Screen" properties:@{}];
     });
-    
+
     it(@"screen with consolidatedPageCalls", ^{
         [integration setSettings:@{
             @"consolidatedPageCalls" : @1
         }];
         [integration screen:[SEGPayloadBuilder screen:@"Home"]];
-        
+
         [verify(mixpanel) track:@"Loaded a Screen" properties:@{@"name": @"Home"}];
     });
 
@@ -117,6 +123,76 @@ describe(@"Mixpanel Integration", ^{
             @"$phone" : @"",
             @"age" : @24
         }];
+    });
+
+
+    it(@"simple group", ^{
+        NSDictionary *groupTraits = @{
+                                      @"groupCity" : @"Cairo",
+                                      @"group_name" : @"mixPanelTest1Group"
+                                      };
+        [integration group:[SEGPayloadBuilder group:@"groupTest1" withTraits:groupTraits]];
+
+        NSDictionary *groupIdentifierTraits = integration.settings[@"groupIdentifierTraits"];
+
+        for (NSString *identifierTrait in groupIdentifierTraits) {
+            NSString *groupIdentifierVal = groupTraits[identifierTrait];
+            if (groupIdentifierVal) {
+                 [verify(mixpanel) getGroup:groupIdentifierVal groupID:@"groupTest1"];
+            }
+
+        }
+
+    });
+
+    it(@"complex group", ^{
+        NSDictionary *groupTraits = @{
+                                      @"groupCity" : @"Alexandria",
+                                      @"name" : @"mixPanelTest1Group",
+                                      @"address": @{@"city": @"Alexandria", @"postalCode": @"22314", @"state":@"Virginia", @"street": @"105 N Union St" },
+                                      @"description": @"Art Center",
+                                      @"avatar" : @"https://gravatar.com/avatar/f8b72def445675a558fe68b1cb651da1?s=400&d=robohash&r=x"
+                                      };
+        [integration group:[SEGPayloadBuilder group:@"groupTest1" withTraits:groupTraits]];
+
+        NSDictionary *groupIdentifierTraits = integration.settings[@"groupIdentifierTraits"];
+
+        for (NSString *identifierTrait in groupIdentifierTraits) {
+            NSString *groupIdentifierVal = groupTraits[identifierTrait];
+            if (groupIdentifierVal) {
+                [verify(mixpanel) getGroup:groupIdentifierVal groupID:@"groupTest1"];
+            }
+
+        }
+
+    });
+
+    it(@"simple group without name", ^{
+        NSDictionary *groupTraits = @{
+                                      @"groupCity" : @"Cairo"
+                                      };
+        [integration group:[SEGPayloadBuilder group:@"groupTest2" withTraits:groupTraits]];
+
+        NSDictionary *groupIdentifierTraits = integration.settings[@"groupIdentifierTraits"];
+
+        for (NSString *identifierTrait in groupIdentifierTraits) {
+            NSString *groupIdentifierVal = groupTraits[identifierTrait];
+             [given([mixpanel getGroup:groupIdentifierVal groupID:@"groupTest2"]) willReturn:nil];
+
+        }
+    });
+
+    it(@"simple group setOnce traits", ^{
+        [given([mixpanel getGroup:@"test1" groupID:@"groupTest2"]) willReturn:mixpanelGroup];
+        NSDictionary *groupTraits = @{
+                                      @"groupCity" : @"Cairo",
+                                      @"groupCount" : @"20",
+                                      @"group_id" : @"test1"
+                                      };
+        [integration group:[SEGPayloadBuilder group:@"groupTest2" withTraits:groupTraits]];
+
+        [verify(mixpanelGroup) setOnce:groupTraits];
+        [verify(mixpanel) getGroup:@"test1" groupID:@"groupTest2"];
     });
 
     it(@"alias", ^{