+++ Pod/Classes/SEGMixpanelIntegration.h
@@ -1,6 +1,11 @@
 #import <Foundation/Foundation.h>
-#import <Analytics/SEGIntegration.h>
 #import <Mixpanel/Mixpanel.h>
+
+#if defined(__has_include) && __has_include(<Analytics/SEGAnalytics.h>)
+#import <Analytics/SEGIntegration.h>
+#else
+#import <Segment/SEGIntegration.h>
+#endif
 
 
 @interface SEGMixpanelIntegration : NSObject <SEGIntegration>